requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
chromadb==0.4.15
sentence-transformers==2.2.2
requests==2.31.0
gitpython==3.1.37
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
watchdog==3.0.0
huggingface-hub==0.16.4  # Add this specific version